requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
firebase-admin
httpx
scikit-learn
numpy
python-dotenv
pydantic[email]